# openapi.yaml
openapi: 3.1.0
info:
  title: LetheCrypt API
  description: API Spec for LetheCrypt, a password manager.
  version: 0.0.0
servers: []
paths:
...
  /pass/{id}:
    get:
      summary: Get passwordfile
      operationId: getPasswordfile
      description: Returns a record of type PasswordFile.
      tags:
        - passcrud
      security:
        - BearerJwt: []
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: The request was successful, and the server has returned the
            requested resource in the response body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordFile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
    patch:
      summary: Patch passwordfile
      operationId: patchPasswordfile
      description: Partially updates a record of type PasswordFile.
      tags:
        - passcrud
      security:
        - BearerJwt: []
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: The request was successful, and the server has returned the
            requested resource in the response body.
          content:
            application/json:
              schema:
                type: "null"
                title: PartialUpdatePasswordFileOk
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
      requestBody:
        required: true
        description: Should have the new content in an object, which will override the
          previously saved content.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordFilePartialUpdate"
    delete:
      summary: Delete passwordfile
      operationId: deletePasswordfile
      description: Deletes a record of type PasswordFile.
      tags:
        - passcrud
      security:
        - BearerJwt: []
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "204":
          description: The request was successful, but there is no content to return in
            the response.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                title: RemovePasswordFileNoContent
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
...
components:
  schemas:
    PasswordFile:
      type: object
      properties:
        author:
          type: string
        content:
          type: string
        createdTimestamp:
          type: string
        name:
          type: string
        passwordID:
          type: number
      required:
        - content
        - createdTimestamp
        - author
        - name
        - passwordID
      description: Represents a password file on the server
...
    PasswordFilePartialUpdate:
      type: object
      properties:
        content:
          type: string
        name:
          type: string
      required: []
      description: Data transfer object for partially updating an existing
        PasswordFile (PATCH operation).
      x-fiddle-dto-info:
        baseSchemaName: PasswordFile
...
  responses:
    BadRequest:
      description: The server could not understand the request due to invalid syntax.
        The client should modify the request and try again.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            required:
              - message
    NotFound:
      description: The server cannot find the requested resource. The endpoint may be
        invalid or the resource may no longer exist.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            required:
              - message
...
    InternalServerErrror:
      description: The server encountered an unexpected condition that prevented it
        from fulfilling the request. Report the issue to the support team if it
        persists.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            required:
              - message
    Unauthorized:
      description: Authentication is required to access the requested resource. The
        client must include the appropriate credentials.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            required:
              - message
  parameters:
    id:
      name: id
      in: path
      description: ""
      schema:
        type: number
      explode: false
      style: simple
      required: true
  securitySchemes:
    BearerJwt:
      type: http
      scheme: Bearer
      bearerFormat: Json Web Token (JWT)
      description: "Jwt Auth: Authenticated requests contain a valid Json Web Token
        (JWT) as part of the `Authorization: Bearer <JWT>` header."
...